backend/requirements.txt
# Core Framework
flask==3.0.2
flask-cors==4.0.0

# Computer Vision & ML (Strict Versioning for Python 3.10.10)
numpy==1.26.4
opencv-python==4.9.0.80
tensorflow==2.15.0
tf-keras==2.15.0
deepface==0.0.83

# Database & Storage
pymongo==4.6.1
supabase==2.3.4

# Utilities
python-dotenv==1.0.1
gunicorn==21.2.0
requests==2.31.0
pillow==10.2.0
